moved query page code to fragment
